Backend/requirements.txt
flask
python-dotenv
bcrypt
pyjwt
psycopg2-binary